Added ability to change block sprite
I added a new block sprite that is much easier to distinguish from the original. The player can choose, in a new options menu, whether to play with the old block sprite or the new one. Their choice will be saved for a file named "options.cfg" and reloaded every time.